moving field types into separate file, accessing widget from qlistwidget
